A solution for an issue involving the breaking of links.
When installing the Ubuntu distro for WSL 2, the users have a choice. They can choose to install a specific version, such as either Ubuntu 20 or Ubuntu 22. They can also choose to simply install the latest LTS version, which is identified as simply "Ubuntu" by the system without a version number but as of now, it's actually "Ubuntu 22" because that's the most recent LTS version currently. It means that the "${WSL_DISTRO_NAME}" variable can point to "Ubuntu" when "Ubuntu 22" is installed. That can break the link created by some commands in the second step, named "usage", of the README[dot]md file for this repository if a "config-Ubuntu" file isn't provided for "Ubuntu 22" in the "linux-msft-wsl-x.xx.xx.x" directories. My system's windows build number is "5.15.90.1", so it's possible for me to confirm the situation for it. In other words, the "config-Ubuntu-20[dot]04" & "config-Ubuntu-22[dot]04" files were already present but a "config-Ubuntu" was needed in the "linux-msft-wsl-5.15.90.1" directory to avoid the breaking of links, so it's being added by me.